search

+++ src/mmdbs/lsi/IndexReaderInterface.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 
 public interface IndexReaderInterface {
     public String[] getTerms() throws Exception;
-    public int[][] getTermDocumentMatrix() throws Exception;
+    public double[][] getTermDocumentMatrix() throws Exception;
 }
